Split the loader into it's own files
